# Peken-Banyumasan-Gate-Masuk-Keluar/openapi-gate.yaml
openapi: 3.1.0
info:
  title: Peken Banyumasan — Gate (Admin & Petugas) API
  version: 2.4.1
  description: |
    API consumed by the Gate admin/petugas panel
    (`Peken-Banyumasan-Gate-Masuk-Keluar/frontend/`).

    **Auth (v2.0 — Supabase):**
    - Login/logout/me/changePassword are **Supabase-direct** — frontend calls
      `supabase.auth.signInWithPassword()` etc. without touching this backend.
    - Gate has **no self-register** — admin and petugas accounts are created by superadmin.
    - `PUT /api/auth/profile` (custom fields only) is **HYBRID**.
    - All other endpoints require a valid **Supabase JWT** in `Authorization: Bearer <token>`.
      Backend verifies using `SUPABASE_JWT_SECRET` and enforces `app_metadata.role IN ('admin', 'petugas')`.
    - Endpoints exclusive to `role: admin` are noted in their summary. Petugas is limited to
      dashboard, visitor scan, and basic event/zone read.

    **Vocabulary (CANONICAL — non-negotiable):**
    - `kolaborator` (NEVER member, kreator)
    - `artisan` (NEVER tenant, umkm)
    - `aktivitas` for admin-moderation story feed (NEVER admin/stories)
    - Path segment `/kolaborator` singular (NEVER `/kolaborators`)
    - Path segment `/artisan` singular (NEVER `/artisans`)

    ---

    **Response Envelope (ALL endpoints except binary file downloads):**
    ```
    Success: { "status": "success", "message": "<optional>", "data": <payload> }
...
    ```
...

    **Authentication:**
...
    does NOT need to implement those endpoints.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/events:
    get:
      operationId: listEvents
      tags: [events-crud]
      summary: List all events (admin sees all statuses)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, berlangsung, selesai]
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEvent
      tags: [events-crud]
      summary: Create new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventBody'
            example:
              nama: Peken Banyumas #8
              tanggal: '2026-08-10'
              tanggal_selesai: '2026-08-12'
              jam_mulai: '09:00'
              jam_selesai: '21:00'
              lokasi: Alun-alun Purwokerto
              status: draft
              kapasitas: 500
              deskripsi: Pameran kreatif terbesar ke-8
              konten_lengkap: '...'
              subsektor: [Kriya, Fashion, Kuliner]
              banner_url: null
              galeri: []
      responses:
        '200':
          description: Created event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/events/{id}:
    get:
      operationId: getEvent
      tags: [events-crud]
      summary: Get event detail (admin sees all statuses)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateEvent
      tags: [events-crud]
      summary: Full update event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventBody'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteEvent
      tags: [events-crud]
      summary: Delete event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Event dihapus
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/events/{id}/status:
    patch:
      operationId: patchEventStatus
      tags: [events-crud]
      summary: Change event status (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, published, berlangsung, selesai]
            example:
              status: published
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          status:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Events ↔ Kolaborator (Admin Assignment) ───────────────────────────────────
  /api/events/{id}/kolaborator:
    get:
      operationId: listEventKolaborator
      tags: [events-kolaborator]
      summary: List kolaborators assigned to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: EventKolaborator list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: assignKolaboratorToEvent
      tags: [events-kolaborator]
      summary: Admin-assign kolaborator to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kolaborator_id, peran]
              properties:
                kolaborator_id:
                  type: string
                peran:
                  type: string
                  enum: [performer, panitia, peserta]
                status_kehadiran:
                  type: string
                  enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
                  default: terdaftar
            example:
              kolaborator_id: kol-0001
              peran: performer
              status_kehadiran: terdaftar
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/events/{id}/kolaborator/{kid}:
    patch:
      operationId: updateEventKolaborator
      tags: [events-kolaborator]
      summary: Update kolaborator assignment in event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: kid
          in: path
          required: true
          schema:
            type: string
          description: EventKolaborator junction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                peran:
                  type: string
                status_kehadiran:
                  type: string
                  enum: [terdaftar, hadir, tidak_hadir, dibatalkan]
            example:
              status_kehadiran: hadir
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventKolaborator'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: removeEventKolaborator
      tags: [events-kolaborator]
      summary: Remove kolaborator from event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: kid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Kolaborator dihapus dari event
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/events/{id}/kolaborator-requests:
    get:
      operationId: listKolaboratorRequests
      tags: [events-kolaborator]
      summary: List pending kolaborator self-join requests (admin only)
      description: Returns requests with status='pending'. Submitted via Kolaborator app.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Pending kolaborator requests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/KolaboratorRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/events/{id}/kolaborator-requests/{rid}:
    patch:
      operationId: respondKolaboratorRequest
      tags: [events-kolaborator]
      summary: Approve or reject a kolaborator self-join request (admin only)
      description: |
        **Side effects on approve:**
        - Insert `event_kolaborator` row with `assigned_by: 'self'`, `status_kehadiran: 'terdaftar'`
        - Emit notification to kolaborator (`type: 'event_invite'`)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: rid
          in: path
          required: true
          schema:
            type: string
          description: KolaboratorRequest ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve, reject]
            example:
              action: approve
      responses:
        '200':
          description: Request responded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Permintaan disetujui
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Events ↔ Artisan (Admin Assignment) ──────────────────────────────────────
  /api/events/{id}/artisan:
    get:
      operationId: listEventArtisan
      tags: [events-artisan]
      summary: List artisans assigned to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: EventArtisan list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: assignArtisanToEvent
      tags: [events-artisan]
      summary: Admin-assign artisan to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artisan_id]
              properties:
                artisan_id:
                  type: string
                stand_id:
                  type: string
                  nullable: true
                  description: Concrete zone stand (validated against the event's zones)
                posisi_event:
                  type: string
                  nullable: true
                  description: Free-text position label (e.g. "Zona A - Stand 5"); falls back to stand_id when omitted
                status_request:
                  type: string
                  enum: [approved]
                  default: approved
            example:
              artisan_id: art-0001
              posisi_event: Zona A - Stand 5
              status_request: approved
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/events/{id}/artisan/{aid}:
    patch:
      operationId: updateEventArtisan
      tags: [events-artisan]
      summary: Update artisan assignment in event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: aid
          in: path
          required: true
          schema:
            type: string
          description: EventArtisan junction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stand_id:
                  type: string
                  nullable: true
                posisi_event:
                  type: string
                  nullable: true
                status_request:
                  type: string
                  enum: [pending, pending_change, approved, rejected]
            example:
              stand_id: B-2
              status_request: approved
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: removeEventArtisan
      tags: [events-artisan]
      summary: Remove artisan from event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: aid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Artisan dihapus dari event
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/events/{id}/artisan-requests:
    get:
      operationId: listArtisanRequests
      tags: [events-artisan]
      summary: List pending/pending_change artisan self-join requests (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Artisan requests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArtisanRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/events/{id}/artisan-requests/{rid}:
    patch:
      operationId: respondArtisanRequest
      tags: [events-artisan]
      summary: Approve or reject initial artisan join request (admin only)
      description: |
        Responds to initial `requestJoin` (step 1 of the artisan-request flow).
        **Side effects on approve:**
        - Insert/update `event_artisan` row with `status_request: 'approved'`, `stand_id` assigned
        - Mark zone stand `occupied: true`
        - Emit notification to artisan (`type: 'artisan_request_approved'`)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: rid
          in: path
          required: true
          schema:
            type: string
          description: ArtisanRequest ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve, reject]
                stand_id:
                  type: string
                  description: Required when action='approve'
                  example: A-3
            example:
              action: approve
              stand_id: A-3
      responses:
        '200':
          description: Request responded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Permintaan artisan disetujui
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/kolaborator/{id}/events:
    get:
      operationId: listKolaboratorEvents
      tags: [kolaborator-management]
      summary: List events a kolaborator has joined (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Events list (event rows the kolaborator has joined)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/KolaboratorEventEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT (RS256). Issued by `supabase.auth.signInWithPassword()` on the frontend.
        Backend verifies using `SUPABASE_JWT_SECRET`.
        Role must be 'admin' OR 'petugas' in `app_metadata.role`.
        Endpoints marked "admin only" additionally reject role 'petugas' with 403.
...
  schemas:
    Envelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
          nullable: true
        data:
          description: Actual payload — shape varies per endpoint
...